requirements.txt
Django
PyPDF2
python-docx
Pillow
pandas
openpyxl
xlrd
python-magic
filetype
requests
anthropic
ollama
python-dotenv
pytest
pytest-django
django-crispy-forms
httpx
pandas
pytest-cov 